split vizModel to a separate file (#5094)
vision.go was super long, and half of it was unrelated to the other half. So, it's now 2 files. This will make it easier to think about fixing RDSK-11093, which is about updating `vision.NewService()` to not use the deprecated constructor.

No changes to functionality are intended: I just took a file and split it in half.
